11.8.8.0

+++ com/baidu/location/f.java
@@ -6,12 +6,14 @@ import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
 import android.os.IBinder;
+import android.os.Process;
+import android.util.Log;
 import com.baidu.android.util.io.ActionJsonData;
-import com.baidu.location.d.j;
+import com.baidu.location.e.l;
 import dalvik.system.DexClassLoader;
 import java.io.File;
 import java.io.RandomAccessFile;
-/* loaded from: classes10.dex */
+/* loaded from: classes20.dex */
 public class f extends Service {
     LLSInterface a = null;
     LLSInterface b = null;
@@ -25,7 +27,7 @@ public class f extends Service {
         int readInt;
         boolean z = false;
         try {
-            File file2 = new File(j.h() + "/grtcfrsa.dat");
+            File file2 = new File(l.j() + "/grtcfrsa.dat");
             if (file2.exists()) {
                 RandomAccessFile randomAccessFile = new RandomAccessFile(file2, "rw");
                 randomAccessFile.seek(200L);
@@ -33,8 +35,8 @@ public class f extends Service {
                     byte[] bArr = new byte[readInt];
                     randomAccessFile.read(bArr, 0, readInt);
                     String str = new String(bArr);
-                    String a = j.a(file, "SHA-256");
-                    if (str != null && a != null && j.b(a, str, "MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCiP7BS5IjEOzrKGR9/Ww9oSDhdX1ir26VOsYjT1T6tk2XumRpkHRwZbrucDcNnvSB4QsqiEJnvTSRi7YMbh2H9sLMkcvHlMV5jAErNvnuskWfcvf7T2mq7EUZI/Hf4oVZhHV0hQJRFVdTcjWI6q2uaaKM3VMh+roDesiE7CR2biQIDAQAB")) {
+                    String a = l.a(file, "SHA-256");
+                    if (str != null && a != null && l.b(a, str, "MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCiP7BS5IjEOzrKGR9/Ww9oSDhdX1ir26VOsYjT1T6tk2XumRpkHRwZbrucDcNnvSB4QsqiEJnvTSRi7YMbh2H9sLMkcvHlMV5jAErNvnuskWfcvf7T2mq7EUZI/Hf4oVZhHV0hQJRFVdTcjWI6q2uaaKM3VMh+roDesiE7CR2biQIDAQAB")) {
                         z = true;
                     }
                 }
@@ -46,7 +48,7 @@ public class f extends Service {
     }
 
     public static float getFrameVersion() {
-        return 7.63f;
+        return 9.02f;
     }
 
     public static String getJarFileName() {
@@ -57,28 +59,39 @@ public class f extends Service {
         return mC;
     }
 
+    public static void setServiceContext(Context context) {
+        mC = context;
+    }
+
     @Override // android.app.Service
     public IBinder onBind(Intent intent) {
-        return this.c.onBind(intent);
+        if (this.c != null) {
+            return this.c.onBind(intent);
+        }
+        return null;
     }
 
     @Override // android.app.Service
     @SuppressLint({"NewApi"})
     public void onCreate() {
+        if (isServing) {
+            Log.d("baidu_location_service", "baidu location service can not start again ...20190306..." + Process.myPid());
+            return;
+        }
         mC = getApplicationContext();
         System.currentTimeMillis();
-        this.b = new com.baidu.location.c.a();
+        this.b = new com.baidu.location.d.a();
         try {
-            File file = new File(j.h() + File.separator + replaceFileName);
-            File file2 = new File(j.h() + File.separator + "app.jar");
+            File file = new File(l.j() + File.separator + replaceFileName);
+            File file2 = new File(l.j() + File.separator + "app.jar");
             if (file.exists()) {
                 if (file2.exists()) {
                     file2.delete();
                 }
                 file.renameTo(file2);
             }
-            if (file2.exists() && a(new File(j.h() + File.separator + "app.jar"))) {
-                this.a = (LLSInterface) new DexClassLoader(j.h() + File.separator + "app.jar", j.h(), null, getClassLoader()).loadClass("com.baidu.serverLoc.LocationService").newInstance();
+            if (file2.exists() && a(new File(l.j() + File.separator + "app.jar"))) {
+                this.a = (LLSInterface) new DexClassLoader(l.j() + File.separator + "app.jar", l.j(), null, getClassLoader()).loadClass("com.baidu.serverLoc.LocationService").newInstance();
             }
         } catch (Exception e) {
             this.a = null;
@@ -97,7 +110,9 @@ public class f extends Service {
     @Override // android.app.Service
     public void onDestroy() {
         isServing = false;
-        this.c.onDestroy();
+        if (this.c != null) {
+            this.c.onDestroy();
+        }
         if (isStartedServing) {
             stopForeground(true);
         }
@@ -124,11 +139,13 @@ public class f extends Service {
 
     @Override // android.app.Service
     public void onTaskRemoved(Intent intent) {
-        this.c.onTaskRemoved(intent);
+        if (this.c != null) {
+            this.c.onTaskRemoved(intent);
+        }
     }
 
     @Override // android.app.Service
     public boolean onUnbind(Intent intent) {
-        return this.c.onUnBind(intent);
+        return false;
     }
 }
